Changed  quorum threshold, dynamic notifications
Quorum threshold is now 3% of all delegates at runtime. This might cause issues with proposals jumping back and forth across the threshold, but we will have to see.

There's now a quorum.json file that stores channel IDs and messages for having notifications in more than just the one channel as well.